udemy/master_python/requirements.txt
mysql-connector-python==8.3.0
python-decouple==3.8
python-dotenv==1.0.1
